requirements.txt
Flask==2.3.2
pandas==2.2.3
openpyxl==3.1.5
gunicorn==20.1.0
psycopg2==2.9.10
SQLAlchemy==2.0.36
reportlab==4.2.5
python-dotenv==1.0.1
Werkzeug>=2.0.0
matplotlib==3.7.4